third_party/WebKit/LayoutTests/shadow-dom/event-deep-path-empty.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id=target></div>
<script>
test(() => {
  const ev = new Event("sample");
  var dispatched = false;
  target.addEventListener("sample", (e) => {
    assert_true(e.deepPath().length > 0);
    dispatched = true;
  });
  assert_array_equals(ev.deepPath(), []);
  target.dispatchEvent(ev);
  assert_true(dispatched);
  assert_array_equals(ev.deepPath(), []);
}, "Event.deepPath() should return an empty array before/after dispatching an event");
</script>
